Add files via upload
Basically, when the user selects a place from the dropdown list, the handlePlaceChange function updates the selectedPlace state variable with the value of the selected place.
That triggers the useEffect hook that depends on the selectedPlace variable. The hook then makes an HTTP GET request to /users?place=${selectedPlace} using the axios, passing the selected place as a parameter.
The response data is used to update the users state variable using the setUsers. The component then displays the list of users in the selected place.

Not completely sure that it works, tho.